% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{alpha_cov}
\alias{alpha_cov}
\title{Calculate covariate-free and covariate-dependent Cronbach's alpha}
\usage{
alpha_cov(data, xx, zz = NULL, use = "pairwise")
}
\arguments{
\item{data}{A data frame or matrix containing both the variables as well as
the covariate(s). This can also be a covariance matrix.}

\item{xx}{Numeric. A vector of indices to specify the columns in \code{data}
that pertain to the target variables.}

\item{zz}{Numeric. A vector of indices to specify the columns in \code{data}
that pertain to the covariate(s).}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\value{
Returns a named vector with covariate-free, covariate-dependent, and
  total Cronbach's alpha (unstandardized).
}
\description{
If a covariate (or possibly a set of covariates) is partialled out from a set
of target variables, their corresponding coefficient alpha can be decomposed 
into a part that is dependent on that covariate and a part that is 
independent ("free") from that covariate (and the sum equals ordinary
coefficient alpha).
}
\references{
Bentler, P. (2015). \emph{Covariate-free and covariate-dependent
  reliability}. Manuscript submitted for publication. Retrieved from
  \url{http://escholarship.org/uc/item/6v61v1pk}
}
\seealso{
Cronbach's alpha in the \strong{psych} package: \code{\link[psych]{alpha}}
}

