% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{barplot_main}
\alias{barplot_main}
\title{Make a barplot and include the call to \link{barplot} in the main title}
\usage{
barplot_main(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link{barplot}}}
}
\description{
Internal helper function
}

