% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_replicate-simulation-for-roc.R
\name{performance_style}
\alias{performance_style}
\title{Rearrange Simulation Results in Order to Plot Them}
\usage{
performance_style(x, dat)
}
\arguments{
\item{x}{Character. One of \code{"TPR"}, \code{"FPR"}, \code{"ACC"}, or \code{"PREC"}.}

\item{dat}{List. The output from \code{\link{predict_style}}.}
}
\value{
An S4 object of the ROCR-class performance.
}
\description{
This function takes as input the output from \code{\link{predict_style}} and
rearranges this output so that it can subsequently be used by the plotting
function of the \strong{ROCR} package (\code{\link[ROCR]{plot.performance}}). Conversely to traditional
ROC-analyses, which employ two measures (e.g., TPR and FPR), this function
returns output to plot only one of these measures (e.g., TPR).
}
\examples{
\dontrun{
res_1 <- predict_style(reps = 5, style = "ARS")
res_2 <- performance_style("ACC", res_1)
ROCR::plot(res_2)
ROCR::plot(res_2, avg="vertical", spread.estimate="stddev",
           show.spread.at=seq(.1, .9, length=9))
}
}
\seealso{
The package \strong{ROCR} and its \code{\link[ROCR]{performance-class}}.
}

