% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_utils.R
\name{plot_x}
\alias{plot_x}
\title{Plot a default scatterplot and add simple slope regression lines in order to
visualize interaction effects}
\usage{
plot_x(data, model, y, x, xq = NULL, lvl = NULL, N = NULL, ylab = y,
  xlab = x[1], lwd = 2, lty = 1, col = "black", pch = 20,
  alpha = 0.05, col.s = NULL, xlim = range(data[, x[1]]),
  ylim = range(data[, y]), seg.len = 0.5, cex.legend = 0.8,
  plot.legend = TRUE, at.x = NULL, at.y = NULL, labels.x = TRUE,
  labels.y = TRUE, cex.x.axis = 1, cex.y.axis = 1, grid = TRUE,
  grid.h = NULL, padj.x = NA, padj.y = NA)
}
\arguments{
\item{data}{A data frame, the same one that was used to fit the model.}

\item{model}{An object of class (\code{\link[stats]{lm}}). Make sure to use 
the \code{data} argument of \code{\link[stats]{lm}} (i.e., do not use 
something like \code{lm(d$y ~ d$x1*d$x2)}).}

\item{y}{Character. The dependent variable in the \code{model}.}

\item{x}{Character vector. The independent variable(s) in the \code{model}
for which the interaction effects should be plotted. If this is of length >
1, the first element will be used as the x-variable of the plot and the
second element will be used for different simple slopes (and the third
variable will be used for different plot panels).}

\item{xq}{Character. Optional term in your model, namely, the quadratic 
effect of your first independent variable.}

\item{lvl}{List of numeric vectors. The list must be of length 1 with two 
predictors and of length 2 with three predictors. Each vector contains the 
values of the corresponding predictor at which the plot is to be drawn.
Defaults to three values, namely \eqn{M - 1SD}, \eqn{M}, and \eqn{M + 1SD}.}

\item{N}{Numeric. A number of data points to plot; possibly smaller than 
\code{nrow(data)}.}

\item{ylab}{Y axis label, same font attributes as \code{xlab}.}

\item{xlab}{X axis label using font and character expansion
    \code{par("font.lab")} and color \code{par("col.lab")}.}

\item{lwd}{Numeric. The line width(s) for the line(s) (see
\code{\link[graphics]{par}}).}

\item{lty}{Numeric. The line type(s) for the line(s) (see
\code{\link[graphics]{par}}).}

\item{col}{Character. The color of the points.}

\item{pch}{the plotting symbols appearing in the legend, as
    numeric vector or a vector of 1-character strings (see
    \code{\link{points}}). Unlike \code{points}, this can all be
    specified as a single multi-character string.  \emph{Must} be
    specified for symbol drawing.}

\item{alpha}{Numeric. An alpha transparency value (as an opacity, so 0 means 
fully transparent and 1 means opaque; see \code{\link[grDevices]{rgb}}).}

\item{col.s}{Character vector. The color(s) of the slope(s). Must be of 
length 1 for one predictor and equal to \code{length(lvl[[1]])} for 2 
predictors.}

\item{xlim, ylim}{Numeric vectors of length 2, giving the x and y coordinates ranges.}

\item{seg.len}{the length of lines drawn to illustrate \code{lty}
    and/or \code{lwd} (in units of character widths).}

\item{cex.legend}{Numeric. A value giving the amount by which the legend 
should be magnified (see \code{\link[graphics]{par}}).}

\item{plot.legend}{Logical. Should a legend be plotted.}

\item{at.x, at.y}{Numeric. The points at which tick-marks for the x- and the 
y-axis, respectively, are to be drawn (see \code{\link[graphics]{axis}}).}

\item{labels.x, labels.y}{This can either be a logical value specifying
whether (numerical) annotations are to be made at the tickmarks of the x-
and y-axsi, respetively, or a character or expression vector of labels to
be placed at the tickpoints. If this is not logical, \code{at.x} and/or 
\code{at.y} should also be supplied and of the same length (see 
\code{\link[graphics]{axis}}).}

\item{cex.x.axis, cex.y.axis}{The magnification to be used for axis annotation
relative to the current setting of cex (see \code{\link[graphics]{par}}).}

\item{grid}{Logical. Should a grid be plotted.}

\item{grid.h}{Numeric. Values of the y-axis at which the horizontal grid 
lines should be plotted.}

\item{padj.x, padj.y}{Numeric. Adjustment for each tick label perpendicular to
the reading direction. For labels parallel to the axes, padj = 0 means 
right or top alignment, and padj = 1 means left or bottom alignment (see 
\code{\link[graphics]{axis}}).}
}
\description{
This function may help to understand and interpret interaction effects
observed in a linear model (\code{\link[stats]{lm}}) by means of simple
slopes.
}
\seealso{
\code{\link[graphics]{par}}, \code{\link[graphics]{plot}}, and
  \code{\link{legend}}
}

