% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_replicate-simulation.R
\name{rep_style_rel}
\alias{rep_style_rel}
\title{Replicate a Style Simulation and Investigate Effect on Cronbach's Alpha}
\usage{
rep_style_rel(reps = 1000, n = c(100, 1000), items = c(5, 10),
  categ = c(3, 7), ndimc = 1, style = NULL, reversed = c(0, 0.5),
  mu.s = c(-1, 1), var.s = c(0, 1), df = 10, sig = NULL, emp = TRUE,
  ...)
}
\arguments{
\item{reps}{Numeric, the desired number of replications.}

\item{n}{Numeric, the number of persons. If of length one, it's fixed to the 
provided value. If of length two, it's sampled from a uniform distribution 
using the two values as lower and upper limits, respectively.}

\item{items}{Numeric, the number of items. If of length one, it's fixed to
the provided value. If of length two, it's sampled from a uniform
distribution using the two values as lower and upper limits, respectively.}

\item{categ}{Numeric, the number of categories per item. If of length one,
it's fixed to the provided value. If of length two, it's sampled from a
uniform distribution using the two values as lower and upper limits,
respectively..}

\item{ndimc}{Numeric. Desired number of content-related latent variables (irrespective
of number of style-related latent variables).}

\item{style}{Parameter to specify which response style(s) influence the data,
can be either numeric or character. Users may choose one or more among
\code{"ERS1"} (e.g., 1 / 0 / 0 / 0 / 1), \code{"ERS2"} (e.g., 2 / 1 /
0 / 1 / 2), \code{"ARS"} (e.g., 0 / 0 / 0 / 1 / 1), \code{"ADRS"}
(e.g, -1 / -1 / 0 / 1 / 1), and \code{"MRS"} (e.g., 0 / 0 / 1 / 0 /
0). Alternatively, a user-specified vector of weights can be employed. Can
also be \code{NULL} indicating complete abscence of response styles.}

\item{reversed}{Numeric, the number of reverse-coded items. If of length one,
it's fixed to the provided value. If of length two, it's sampled from a
uniform distribution using the two values as lower and upper limits,
respectively.}

\item{mu.s}{Numeric, the response style mean. If of length one, it's fixed to
the provided value. If of length two, it's sampled from a uniform
distribution using the two values as lower and upper limits, respectively.}

\item{var.s}{Numeric, the response style variance. If of length one, it's
fixed to the provided value. If of length two, it's sampled from a uniform
distribution using the two values as lower and upper limits, respectively.}

\item{df}{Numeric. The df-parameter of the Wishart distribution from which
the covariance is drawn.}

\item{sig}{Numeric matrix. The variance-covariance matrix of the multivariate
distribution of thetas. If non-NULL, this overrides \code{var.s}.}

\item{emp}{Logical. If true, \code{mu.s} and \code{var.s}/\code{sig} specify the empirical not population mean and covariance matrix.}

\item{...}{Other parameters passed to \code{\link{sim_style_data}}.}
}
\value{
Returns a matrix of length \code{reps} with the following columns:
\item{bias}{\code{alpha} minus \code{true}}
\item{true}{Response style-free alpha}
\item{alpha}{Observed coefficient alpha}
\item{dep}{Response style-dependent alpha, equal to bias}
\item{ }{Further columns contain the input parameters such as the number of categories}
}
\description{
This function replicates \code{\link{sim_style_data}} and returns the observed
coefficient alpha as well as the response style-free alpha for every
replication sample.
}
\seealso{
The replicated function \code{\link{sim_style_data}}, covariate-free alpha \code{\link{alpha_cov}}
}

