\docType{package}
\name{stylesim}
\alias{stylesim}
\alias{stylesim-package}
\title{Simulate (and Analyse) Data Distorted by Response Styles}
\description{
The most important function is \code{sim_style_data()},
which allows to simulate data distorted by response styles
}
\author{
Hansjoerg Plieninger
}

