% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_replicate-simulation-for-roc.R
\name{predict_style}
\alias{predict_style}
\title{Simulate Data Distorted by Response Style and Calculate, for Example, the
True Positive Rate for Subsequent ROC Analyses}
\usage{
predict_style(reps = 100, n = 200, items = 10, categ = 5, ndimc = 1,
  style = NULL, reversed = 1/3, mu.s = 0, var.s, sig = NULL,
  emp = TRUE, cutoff, ...)
}
\arguments{
\item{reps}{Desired number of replications}

\item{n}{Numeric. Desired number of observations/persons.}

\item{items}{Numeric. Desired number of items/stimuli. If \code{ndimc} > 1, this
specifies the number of items per dimension.}

\item{categ}{Numeric. Desired number of categories per item.}

\item{ndimc}{Numeric. Desired number of content-related latent variables (irrespective
of number of style-related latent variables).}

\item{style}{Parameter to specify which response style(s) influence the data,
can be either numeric or character. Users may choose one or more among
\code{"ERS1"} (e.g., 1 / 0 / 0 / 0 / 1), \code{"ERS2"} (e.g., 2 / 1 /
0 / 1 / 2), \code{"ARS"} (e.g., 0 / 0 / 0 / 1 / 1), \code{"ADRS"}
(e.g, -1 / -1 / 0 / 1 / 1), and \code{"MRS"} (e.g., 0 / 0 / 1 / 0 /
0). Alternatively, a user-specified vector of weights can be employed. Can
also be \code{NULL} indicating complete abscence of response styles.}

\item{reversed}{Numeric. Indicates the number of reverse-coded items. Can be either a
ratio between 0 and 1 indicating the ratio of reverse-coded items or an
integer between 0 and \code{items}) indicating the number of reverse-coded
items.}

\item{mu.s}{Numeric. The mean(s) of the response style distribution.}

\item{var.s}{Numeric. The variance(s) of the response style distribution.}

\item{sig}{Numeric matrix. The variance-covariance matrix of the multivariate
distribution of thetas. If non-NULL, this overrides \code{var.s}.}

\item{emp}{Logical. If true, \code{mu.s} and \code{var.s}/\code{sig} specify the empirical not population mean and covariance matrix.}

\item{cutoff}{Optional numeric vector. Desired cutoff- (alpha-) values at which
to calculate TPR, FPR, etc; probably a sequence, for example from .95 to
.05 (must be within the open interval of (0, 1))}

\item{...}{Other parameters passed to \code{\link{sim_style_data}}.}
}
\value{
\item{alpha}{The cutoff-/alpha-values}
  \item{TPR}{True positive rate}
  \item{FPR}{False positive rate}
  \item{PREC}{Precision}
  \item{ACC}{Accuracy}
  \item{}{Further elements containing the input parameters}
}
\description{
This function repeatedly simulates data that are distorted by response style. It makes
use of the scale scores \eqn{\bar{X}}---which are influenced by both the attribute and
response style---as well as the true person parameters \eqn{\theta_1}, that is, the
persons' standing on the attribute, which is unaffected by response style.
These two scores are used to classify each individual at certain cutoff
values as either above (i.e., positive) or below (i.e., negative) the cutoff.
Subsequently, the true positive rate (TPR), the false positive rate (FPR),
the precision, and the accuracy are calculated at every cutoff value.
}
\seealso{
The replicated function \code{\link{sim_style_data}}.
}

