% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_replicate-simulation.R
\name{rep_style_cor}
\alias{rep_style_cor}
\title{Replicate a Style Simulation and Investigate Effect on Correlation of Two Scale Scores}
\usage{
rep_style_cor(reps = 1000, n = c(100, 1000), items = c(5, 10),
  categ = c(3, 7), ndimc = 2, style = NULL, reversed = c(0, 0.5),
  mu.s = c(-1, 1), var.s = c(0, 1), df = 10, sig = NULL, emp = TRUE,
  ...)
}
\arguments{
\item{reps}{Numeric, the desired number of replications.}

\item{n}{Numeric. Desired number of observations/persons.}

\item{items}{Numeric. Desired number of items/stimuli. If \code{ndimc} > 1, this
specifies the number of items per dimension.}

\item{categ}{Numeric. Desired number of categories per item.}

\item{ndimc}{Numeric. Desired number of content-related latent variables (irrespective
of number of style-related latent variables).}

\item{style}{Parameter to specify which response style(s) influence the data,
can be either numeric or character. Users may choose one or more among
\code{"ERS1"} (e.g., 1 / 0 / 0 / 0 / 1), \code{"ERS2"} (e.g., 2 / 1 /
0 / 1 / 2), \code{"ARS"} (e.g., 0 / 0 / 0 / 1 / 1), \code{"ADRS"}
(e.g, -1 / -1 / 0 / 1 / 1), and \code{"MRS"} (e.g., 0 / 0 / 1 / 0 /
0). Alternatively, a user-specified vector of weights can be employed. Can
also be \code{NULL} indicating complete abscence of response styles.}

\item{reversed}{Numeric. Indicates the number of reverse-coded items. Can be either a
ratio between 0 and 1 indicating the ratio of reverse-coded items or an
integer between 0 and \code{items}) indicating the number of reverse-coded
items.}

\item{mu.s}{Numeric. The mean(s) of the response style distribution.}

\item{var.s}{Numeric. The variance(s) of the response style distribution.}

\item{df}{Numeric. The df-parameter of the Wishart distribution from which
the covariance is drawn.}

\item{sig}{Numeric matrix. The variance-covariance matrix of the multivariate
distribution of thetas. If non-NULL, this overrides \code{var.s}.}

\item{emp}{Logical. If true, \code{mu.s} and \code{var.s}/\code{sig} specify the empirical not population mean and covariance matrix.}

\item{...}{Other parameters passed to \code{\link{sim_style_data}}.}
}
\value{
Returns a matrix of length \code{reps} with the following columns:
\item{bias}{\code{obs} minus \code{partial}}
\item{partial}{The partial correlation \eqn{r_{\bar{X}_1, \bar{X}_2 \cdot \theta_{\textrm{RS}}}}}
\item{obs}{The observed correlation}
\item{ }{Further columns contain the input parameters such as the number of categories}
}
\description{
This function replicates \code{\link{sim_style_data}} and calculates a scale score
(i.e., the mean across all items) for every person for every dimension. It
mainly returns two values for every replication sample, namely, the observed
correlation of the two scales as well as their partial correlation
controlling for response style.
}
\seealso{
\code{\link{sim_style_data}}
}

