% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylesim-package.r
\docType{package}
\name{stylesim}
\alias{stylesim}
\alias{stylesim-package}
\title{Simulate (and Analyse) Data Distorted by Response Styles}
\description{
The most important function is \code{\link{sim_style_data}}, which allows to
simulate data distorted by response styles.
}
\examples{
\dontrun{
res_1 <- sim_style_data(n = 123, items = 12, categ = 5, style = "ARS", 
                        reversed = 6)
}

}
\author{
Hansjoerg Plieninger
}
\references{
Plieninger, H. (in press). Mountain or molehill: A simulation study on the impact of reponse styles. \emph{Educational and Psychological Measurement}.
}

